move rescue for TypeError from conditional into method body
I'm moving the rescue statement from the coditional into the method body
of constant_name_from_node. If we look at the code we notice that we
have a couple of constant_name method calls that can throw this
exception. By changing the scope we are able to prevent unintended
TypeErrors to crash the program.